+++ backend/springboot/src/main/java/com/workmaite/domain/auth/repository/RefreshTokenRepository.java
@@ -5,13 +5,13 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 
   Optional<RefreshToken> findByTokenHash(String tokenHash);
 
   void deleteByTokenHash(String tokenHash);
 
-  void deleteByUserId(Integer userId);
+  void deleteByUserId(Long userId);
 
-  void deleteByUserIdAndExpiresAtBefore(Integer userId, LocalDateTime now);
+  void deleteByUserIdAndExpiresAtBefore(Long userId, LocalDateTime now);
 }
