+++ backend/springboot/src/main/java/com/workmaite/domain/chat/repository/ChatMessageRepository.java
@@ -5,12 +5,12 @@ import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Integer> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
   List<ChatMessage> findByThreadIdOrderByCreatedAtAsc(String threadId);
 
   // keyset 페이지네이션 (P8-2): id 내림차순으로 beforeId 이전 메시지 limit개 (호출측에서 역순 정렬)
   List<ChatMessage> findByThreadIdAndIdLessThanOrderByIdDesc(
-      String threadId, Integer beforeId, Pageable pageable);
+      String threadId, Long beforeId, Pageable pageable);
 
   List<ChatMessage> findByThreadIdOrderByIdDesc(String threadId, Pageable pageable);
