+++ backend/springboot/src/main/java/com/workmaite/domain/minutes/repository/MinutesRepository.java
@@ -4,11 +4,11 @@ import com.workmaite.domain.minutes.entity.Minutes;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MinutesRepository extends JpaRepository<Minutes, Integer> {
+public interface MinutesRepository extends JpaRepository<Minutes, Long> {
 
-  Optional<Minutes> findBySessionId(Integer sessionId);
+  Optional<Minutes> findBySessionId(Long sessionId);
 
-  boolean existsBySessionId(Integer sessionId);
+  boolean existsBySessionId(Long sessionId);
 
-  void deleteBySessionId(Integer sessionId);
+  void deleteBySessionId(Long sessionId);
 }
