+++ backend/springboot/src/main/java/com/workmaite/domain/reports/dto/ReportSubmitRequest.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class ReportSubmitRequest {
 
   @NotNull(message = "업로더 ID를 입력해주세요.")
-  private Integer uploadId;
+  private Long uploadId;
 
   @NotBlank(message = "제출 부서를 입력해주세요.")
   private String submitterDepartment;
