+++ backend/springboot/src/main/java/com/workmaite/domain/sessions/repository/SessionRepository.java
@@ -8,21 +8,21 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface SessionRepository extends JpaRepository<MeetingSession, Integer> {
+public interface SessionRepository extends JpaRepository<MeetingSession, Long> {
 
-  List<MeetingSession> findByMeetingId(Integer meetingId);
+  List<MeetingSession> findByMeetingId(Long meetingId);
 
-  List<MeetingSession> findByMeetingIdAndStatus(Integer meetingId, SessionStatus status);
+  List<MeetingSession> findByMeetingIdAndStatus(Long meetingId, SessionStatus status);
 
   @Query(
       "SELECT s FROM MeetingSession s WHERE s.meetingId IN (SELECT mm.meetingId FROM MeetingMember mm WHERE mm.userId = :userId) AND s.status = :status ORDER BY s.scheduledAt ASC")
   List<MeetingSession> findUpcomingByUserId(
-      @Param("userId") Integer userId, @Param("status") SessionStatus status);
+      @Param("userId") Long userId, @Param("status") SessionStatus status);
 
   @Query(
       "SELECT s FROM MeetingSession s WHERE s.meetingId IN (SELECT mm.meetingId FROM MeetingMember mm WHERE mm.userId = :userId) AND s.scheduledAt BETWEEN :start AND :end ORDER BY s.scheduledAt ASC")
   List<MeetingSession> findByUserIdAndScheduledAtBetween(
-      @Param("userId") Integer userId,
+      @Param("userId") Long userId,
       @Param("start") LocalDateTime start,
       @Param("end") LocalDateTime end);
 }
